requirements.txt
albumentations==1.3.1
Flask==2.3.3
Flask-PyMongo==2.3.0
Jinja2==3.1.2
MarkupSafe==2.1.3
matplotlib==3.8.2
numpy==1.26.2
pandas==2.1.0
Pillow==10.1.0
pymongo==4.6.0
requests==2.31.0
scikit-image==0.22.0
scikit-learn==1.3.2
seaborn==0.13.0
timm==0.5.4
torch==2.1.1+cu121
torchvision==0.16.1
tornado==6.3.3
tqdm==4.66.1